Add a few tests for TSquareAPI.
The good news is that all but one of the TSquareAPI tests pass, and I'm on the one that doesn't.
The bad news is that our django project has 404 failing test cases. wtf django? I'll try to cut
that down over the next few days, it seems like most of them are configurations that we just
haven't set up yet.