Add universal settings and configure test environment.
In order to run all of the tests in the Django project, run the shell script run-tests.sh.
Although it is specifically a Bash script, it should still work on Windows. All tests
pass except for one that I will tackle soon. There are also two mystery tests that
came default with Django; one is an expected failure and the other is a skip.